fix(ipc): an all-unknown subscribe streams nothing, not everything
Architect round-2 follow-up.

The empty filter meant "everything", which caught two different
requests: `subscribe` with no arguments, which means it, and
`subscribe space_chnage`, which means the opposite. The second
handed back the whole firehose, so a subscription built entirely
of typos read as one working very hard. `data.unknown` explained
it to a client that read the response, but the stream itself
argued otherwise.

Now: arguments given, filter honoured exactly; none given,
everything. A typo-only subscribe is silent, which is exactly as
wrong as the request, and `unknown` says why. No `fail`, so no
`CLIMain` stream-loop change — the architect withdrew that
recommendation once the diagnostic reached the wire.

Tested by ordering rather than a sleep: the emit runs on the main
actor between two round-trips, so a delivered event would be
queued ahead of the second response, and the second read is done
raw so the failure says "this line is an event" instead of a
decoder complaining about a missing `status` key. Reverting the
one-line fix reds it. The log cap is pinned too — six unknown
names, log truncated, wire list whole — because a log-capped,
wire-uncapped asymmetry is what a later reader "simplifies".

`LogSeamRoutingTests` closes the gap the source scan cannot see.
The seven wirings now share one closure, so probing two of them
probes that body: replacing it with `{ _ in }` reds this while
`LogSeamWiringTests` stays green, which is the guard's documented
limit #2 made concrete. It does not enumerate the seams — the
scan keeps the set honest, and closing that half properly wants
runtime discovery over `Mirror`.

Also: the bootstrap comment no longer credits the guard with
preventing divergent closure bodies, which the single closure
prevents by construction and the guard cannot see; `docs/cli.md`
stops claiming the log carries the same names as the wire when it
truncates; `<non-string>` is documented; the guard's prose says
"a `KiwiCore+Bootstrap` file" now that resolution is by prefix;
and `SocketTests` moves to its own file, `IPCTests` having
reached 347 of the 350-line ceiling.
